Run changed notebooks in CI (#409)
## Problem

We need to ensure that notebooks which are heavily featured in our docs
are actually working

## Solution

Implement a workflow to convert a list of specified notebooks into
scripts that can be run in CI. For each notebook (which is a json
document with an `*.ipynb` file extension):
- Create a temporary directory to isolate from other parts of the repo.
Copy the single notebook over to that tmpdir.
- Use the `nbformat` packge to iterate through notebook cells
- If a cell is showing `pip install` steps, include those in a `run.sh`
bash setup script to activate and configure a venv with required
dependencies
- Collect code from other code cells into a python script called
`notebook.py`
- From the workflow, run the `run.sh` script to install needed
dependencies and execute the `notebook.py` script
- Execution without errors considered a success

Since some notebooks are quite slow/lengthy (for example, if computing
embeddings across a large sample dataset) and there are a lot of them
(over 200 notebooks in this repo), we won't run every notebook on every
push because that would be slow/expensive and deter people from making
small improvements. Instead, we will only run those notebooks which have
changed relative to the master branch.

Some notebooks may fail until minor adjustments are made to them to
account for implicit dependencies on things installed in the colab
environment that are not available in our CI environment. For example,
we may need to add a missing a statement like `!pip install pandas`
instead of just using it directly.

## Type of Change

- [x] Infrastructure change (CI configs, etc)