Update docs/langchain-retrieval-augmentation.ipynb (#420)
## Problem

The docs/langchain-retrieval-augmentation.ipynb notebook was pretty far
out of date

## Solution

Updated:
- Dependency versions
- Adjust SDK usage
- Reworked underlying parquet files loaded with `pinecone_datasets` to
avoid the appearance that the notebook has crashed. Previously it loaded
a ton of data from a single parquet file that took several minutes to
download, then immediately trashed a bunch of the rows. Now we loading
only the data used, and I have split the parquet files into smaller
sizes so the user feels as if something is happening.
- Remove unused dependencies
- Begin consuming Pinecone and OpenAI langchain components from
`langchain-pinecone` and `langchain-openai` paakges
- Bumped up to `gpt-4.5-preview` from `gpt-3.5-turbo` for completions