Adding Kumo x Pinecone example for Personalised movie search (#433)
## Problem
On march 27th 2025 Kumo and Pinecone created a joint webinar which
demonstrated how you can use both technologies to build personalization
applications end-to-end by using Kumo to produce embeddings and Pinecone
to power the downstream applications.

Webinar name: Beyond Similarity: Personalizing Recommendations with Kumo
and Pinecone
Link to recording: https://www.youtube.com/watch?v=4XkTarPy9yw
Participants: @arjunpatel7 @BlazStojanovic

We thought that this would be a great addition to the notebooks already
available in this repository. Happy to incorporate any feedback and
improvements required to get this merged.

## Solution
N/A - Adding full notebook

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [x] None of the above: New example notebook of Pinecone usage

## Test Plan
The notebook has been tested and demonstrated to work end to end before
and during the live webinar.