Fix Colab Widgets Error across ALL notebooks (#437)
## Problem

A subset of notebooks saved from Colab were not being loaded on Github
correctly. This was preventing people from seeing the notebook in the
browser, and subsequently, running them at all.
## Solution

Describe the approach you took. Link to any relevant bugs, issues, docs,
or other resources.

Using @jamescalam's fix and guidance from @jseldess, I wrote a script
that removed the widgets field, which resolves the issue without
modifying the behavior of the notebook too much. These extra fields are
created when you work on a notebook in colab, and download/save/upload
it.

See also:
https://gist.github.com/shacharmirkin/7f608c51f5d1c159d5c5791081eb5c6d

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

I inspected two notebooks that were broken, fixed using this script, and
then uploaded them to colab to corroborate. Seems like the fixed worked!