Refresh Semantic Search Docs Example (#444)
## Problem

Describe the purpose of this change. What problem is being solved and
why?

The prior semantic search notebook was outdated and demonstrated many
patterns that are easier to do in Pinecone natively now. Additionally,
the notebook was lengthy and indexed a TON of vectors that took a while
to demonstrate. We need a faster notebook more in line with modern
practices with Pinecone.

## Solution

Describe the approach you took. Link to any relevant bugs, issues, docs,
or other resources.

- Replaced sentence transformers with Pinecone Integrated Inference
- Replaced lengthy dataset with a smaller, more instructive one. The
smaller dataset also can easily be modified for a larger subset if a
user desires
- Better API key handling for colab or local development
- Should be much faster in CI/CD/Github Actions too.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Describe specific steps for validating this change.