Update 04-langchain-chat.ipynb
New LCEL compatible methods.

Also ensured that the old test case which relied on the LLM being less capable was changed to a new test case that works well with a better LLM.

Tried to give an example of when Prompt Templates might be used in place of a dynamic f-strings in functions.