Added HF_TOKEN to github actions (#465)
## Problem

Describe the purpose of this change. What problem is being solved and
why?

Need a HF_TOKEN to appropriately run certain notebooks on CI/CD.

## Solution

Describe the approach you took. Link to any relevant bugs, issues, docs,
or other resources.

Added one to three files in github actions where test notebook script is
run on changes:

- test-notebooks-changed
- test-notebook-all
- action.yml that coordinate the above.

Added as required input and as arguments. 

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Describe specific steps for validating this change.