fix: Updated learn\generation\langchain\handbook\01-langchain-prompt-templates.ipynb to Work with LangChain 0.3 (#451)
Used newer models and new HuggingFaceEndpoint() class (instead of the,
now deprecated, HuggingFaceHub() class. Changed the LLM from
text-davinci-003 to gpt-3.5-turbo, since the former is deprecated. Had
to change the example from prompting the LLM to make jokes in response
to serious questions, to a Markdown formatting task, as newer LLMs are
too good at following orders and the supposedly bad responses were not
actually bad.

## Problem

Describe the purpose of this change. What problem is being solved and
why?

## Solution

Describe the approach you took. Link to any relevant bugs, issues, docs,
or other resources.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Describe specific steps for validating this change.

---------

Co-authored-by: James Briggs <35938317+jamescalam@users.noreply.github.com>