fix: updated lc handbook 01 (#469)
Had to change the example from prompting the LLM to make jokes in
response to serious questions, to a Markdown formatting task, as newer
LLMs are too good at following orders and the supposedly bad responses
were not actually bad.

## Problem

Describe the purpose of this change. What problem is being solved and
why?

## Solution

Describe the approach you took. Link to any relevant bugs, issues, docs,
or other resources.

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

Co-authored-by: Siraj Aizlewood <31565026+Siraj-Aizlewood@users.noreply.github.com>