fix: Updating learn\generation\langchain\handbook\07-langchain-tools.ipynb (#479)
Previously this code assumed the LLM couldn't do maths, which is now an
outdated assumption. Also it assumed the LLM wouldn't use the tools
provided unless given very clear instructions which is no longer the
case. Updated to be more in line with modern LangChain practices,
including using LCEL.

## Problem

Describe the purpose of this change. What problem is being solved and
why?

## Solution

Describe the approach you took. Link to any relevant bugs, issues, docs,
or other resources.

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [x] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Describe specific steps for validating this change.